+++ android/src/main/kotlin/io/piano/flutter/piano_analytics/PianoAnalyticsPlugin.kt
@@ -13,6 +13,8 @@ import io.flutter.plugin.common.StandardMethodCodec
 import io.piano.android.analytics.Configuration
 import io.piano.android.analytics.PianoAnalytics
 import io.piano.android.analytics.model.Event
+import io.piano.android.analytics.model.PrivacyMode
+import io.piano.android.analytics.model.PrivacyStorageFeature
 import io.piano.android.analytics.model.Property
 import io.piano.android.analytics.model.PropertyName
 import io.piano.android.analytics.model.VisitorIDType
@@ -37,28 +39,22 @@ class PianoAnalyticsPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
 
     private lateinit var channel: MethodChannel
 
-    private val consentsPlugin = PianoConsentsPlugin()
-
     private var context: WeakReference<Context?> = WeakReference(null)
 
     override fun onAttachedToActivity(binding: ActivityPluginBinding) {
         context = WeakReference(binding.activity)
-        consentsPlugin.onAttachedToActivity(binding)
     }
 
     override fun onDetachedFromActivityForConfigChanges() {
         context = WeakReference(null)
-        consentsPlugin.onDetachedFromActivityForConfigChanges()
     }
 
     override fun onReattachedToActivityForConfigChanges(binding: ActivityPluginBinding) {
         context = WeakReference(binding.activity)
-        consentsPlugin.onReattachedToActivityForConfigChanges(binding)
     }
 
     override fun onDetachedFromActivity() {
         context = WeakReference(null)
-        consentsPlugin.onDetachedFromActivity()
     }
 
     override fun onAttachedToEngine(flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
@@ -68,14 +64,10 @@ class PianoAnalyticsPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
             StandardMethodCodec(Codec())
         )
         channel.setMethodCallHandler(this)
-
-        consentsPlugin
-        consentsPlugin.onAttachedToEngine(flutterPluginBinding)
     }
 
     override fun onDetachedFromEngine(binding: FlutterPlugin.FlutterPluginBinding) {
         channel.setMethodCallHandler(null)
-        consentsPlugin.onAttachedToEngine(binding)
     }
 
     override fun onMethodCall(call: MethodCall, result: Result) {
@@ -83,6 +75,8 @@ class PianoAnalyticsPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
             when (call.method) {
                 "init" -> handleInit(call)
                 "send" -> handleSend(call)
+                "privacyIncludeStorageFeatures" -> handlePrivacyIncludeStorageFeatures(call)
+                "privacyExcludeStorageFeatures" -> handlePrivacyExcludeStorageFeatures(call)
                 else -> error("Unknown method")
             }
             result.success(null)
@@ -109,14 +103,41 @@ class PianoAnalyticsPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
             Event.Builder(
                 name = name,
                 properties = (it["data"] as? Map<*, *>)?.map { property ->
+                    val propertyName = property.key as? String ?: error("Undefined property name of event \"$name\"")
+                    val value = property.value as? Map<*, *> ?: error("Undefined property value of event \"$propertyName\"")
+                    val propertyValue = value["value"] ?: error("Undefined property value of event \"$propertyName\"")
+                    val forceType = (value["forceType"] as? String)?.let { type ->
+                        Property.Type.entries.firstOrNull { t -> t.prefix == type }
+                    }
                     getProperty(
-                        property.key as? String ?: error("Undefined property name of event \"$name\""),
-                        property.value
+                        propertyName,
+                        propertyValue,
+                        forceType
                     )
                 }?.toMutableSet() ?: mutableSetOf()
             ).build()
         }.toTypedArray()
         PianoAnalytics.getInstance().sendEvents(*events)
+    }
+
+    private fun handlePrivacyIncludeStorageFeatures(call: MethodCall) {
+        val features = call.arg<List<String>>("features").map { feature ->
+            PrivacyStorageFeature.entries.first { it.name == feature }
+        }
+        call.arg<List<String>>("modes").forEach {
+            val mode = getPrivacyMode(it)
+            mode.allowedStorageFeatures += features
+        }
+    }
+
+    private fun handlePrivacyExcludeStorageFeatures(call: MethodCall) {
+        val features = call.arg<List<String>>("features").map { feature ->
+            PrivacyStorageFeature.entries.first { it.name == feature }
+        }
+        call.arg<List<String>>("modes").forEach {
+            val mode = getPrivacyMode(it)
+            mode.allowedStorageFeatures -= features.toSet()
+        }
     }
 
     companion object {
@@ -135,22 +156,33 @@ class PianoAnalyticsPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
         }
 
         @JvmStatic
-        private fun getProperty(name: String, value: Any?) = when (value) {
-            is Boolean -> Property(PropertyName(name), value)
-            is Int -> Property(PropertyName(name), value)
-            is Long -> Property(PropertyName(name), value)
-            is Double -> Property(PropertyName(name), value)
-            is String -> Property(PropertyName(name), value)
+        private fun getProperty(name: String, value: Any?, forceType: Property.Type? = null) = when (value) {
+            is Boolean -> Property(PropertyName(name), value, forceType)
+            is Int -> Property(PropertyName(name), value, forceType)
+            is Long -> Property(PropertyName(name), value, forceType)
+            is Double -> Property(PropertyName(name), value, forceType)
+            is String -> Property(PropertyName(name), value, forceType)
             is Date -> Property(PropertyName(name), value)
             is List<*> -> {
                 when (value.first()) {
-                    is Int -> Property(PropertyName(name), value.filterIsInstance<Int>().toTypedArray())
-                    is Double -> Property(PropertyName(name), value.filterIsInstance<Double>().toTypedArray())
-                    is String -> Property(PropertyName(name), value.filterIsInstance<String>().toTypedArray())
+                    is Int -> Property(PropertyName(name), value.filterIsInstance<Int>().toTypedArray(), forceType)
+                    is Double -> Property(PropertyName(name), value.filterIsInstance<Double>().toTypedArray(), forceType)
+                    is String -> Property(PropertyName(name), value.filterIsInstance<String>().toTypedArray(), forceType)
                     else -> { error("Invalid array value type of property \"$name\"") }
                 }
             }
             else -> error("Invalid type of property \"$name\"")
+        }
+
+        @JvmStatic
+        private fun getPrivacyMode(name: String) = when(name) {
+            "opt-in" -> PrivacyMode.OPTIN
+            "opt-out" -> PrivacyMode.OPTOUT
+            "exempt" -> PrivacyMode.EXEMPT
+            "custom" -> PrivacyMode.CUSTOM
+            "no-consent" -> PrivacyMode.NO_CONSENT
+            "no-storage" -> PrivacyMode.NO_STORAGE
+            else -> error("Invalid privacy mode \"$name\"")
         }
     }
 }