+++ ios/Classes/PianoAnalyticsPlugin.swift
@@ -13,7 +13,7 @@ fileprivate class Reader: FlutterStandardReader {
         case dateType:
             var value: Int64 = 0
             readBytes(&value, length: 8)
-            return dateFormatter.string(from: Date(timeIntervalSince1970: TimeInterval(value / 1000 )))
+            return Date(timeIntervalSince1970: TimeInterval(value / 1000 ))
         default:
             return super.readValue(ofType: type)
         }
@@ -36,8 +36,6 @@ public class PianoAnalyticsPlugin: NSObject, FlutterPlugin {
         )
         let instance = PianoAnalyticsPlugin()
         registrar.addMethodCallDelegate(instance, channel: channel)
-        
-        PianoConsentsPlugin.register(with: registrar)
     }
 
     public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
@@ -82,7 +80,16 @@ public class PianoAnalyticsPlugin: NSObject, FlutterPlugin {
                 guard let name = event["name"] as? String else {
                     throw PluginError.message("Undefined event name")
                 }
-                return Event(name, data: event["data"] as? [String:Any] ?? [:])
+                let properties = try (event["data"] as? [String:Any])?.map { property in
+                    guard let value = property.value as? [String:Any] else {
+                        throw PluginError.message("Undefined property value of event \"\(property.key)\"")
+                    }
+                    guard let propertyValue = value["value"] else {
+                        throw PluginError.message("Undefined property value of event \"\(property.key)\"")
+                    }
+                    return try Property.from(property.key, propertyValue, forceType: Property.forceType(value["forceType"] as? String))
+                }
+                return Event(name, properties: Set(properties ?? []))
             }
         )
     }
@@ -94,6 +101,61 @@ fileprivate extension VisitorIdType {
         switch name {
         case "ADID": return VisitorIdType.IDFA
         default: return VisitorIdType.UUID
+        }
+    }
+}
+
+fileprivate extension Property {
+    
+    static func forceType(_ name: String?) throws -> PA.PropertyType? {
+        guard let name else {
+            return nil
+        }
+        
+        switch name {
+        case "b":
+            return .bool
+        case "n":
+            return .int
+        case "f":
+            return .float
+        case "s":
+            return .string
+        case "d":
+            return .date
+        case "a:n":
+            return .intArray
+        case "a:f":
+            return .floatArray
+        case "a:s":
+            return .stringArray
+        default:
+            throw PluginError.message("Undefined force type \"\(name)\"")
+        }
+    }
+    
+    static func from(_ name: String, _ value: Any, forceType: PA.PropertyType? = nil) throws -> Property {
+        switch value {
+        case let v as Bool:
+            return try Property(name, v, forceType: forceType)
+        case let v as Int:
+            return try Property(name, v, forceType: forceType)
+        case let v as Int64:
+            return try Property(name, v, forceType: forceType)
+        case let v as Double:
+            return try Property(name, v, forceType: forceType)
+        case let v as String:
+            return try Property(name, v, forceType: forceType)
+        case let v as Date:
+            return try Property(name, v)
+        case let v as [Int]:
+            return try Property(name, v, forceType: forceType)
+        case let v as [Double]:
+            return try Property(name, v, forceType: forceType)
+        case let v as [String]:
+            return try Property(name, v, forceType: forceType)
+        default:
+            throw PluginError.message("Undefined type for event property \"\(name)\"")
         }
     }
 }