+++ lib/piano_analytics.dart
@@ -1,50 +1,62 @@
 import 'package:flutter/services.dart';
-import 'package:piano_analytics/piano_consents.dart';
-
-enum PianoAnalyticsVisitorIDType {
-  uuid("UUID"),
-  adid("ADID");
-
-  final String value;
-
-  const PianoAnalyticsVisitorIDType(this.value);
-}
+import 'package:piano_analytics/enums.dart';
 
 class Property {
   final String _name;
   final dynamic _value;
+  final PropertyType? _forceType;
 
-  Property.bool({required String name, required bool value})
+  Property.bool(
+      {required String name, required bool value, PropertyType? forceType})
       : _name = name,
-        _value = value;
+        _value = value,
+        _forceType = forceType;
 
-  Property.int({required String name, required int value})
+  Property.int(
+      {required String name, required int value, PropertyType? forceType})
       : _name = name,
-        _value = value;
+        _value = value,
+        _forceType = forceType;
 
-  Property.double({required String name, required double value})
+  Property.double(
+      {required String name, required double value, PropertyType? forceType})
       : _name = name,
-        _value = value;
+        _value = value,
+        _forceType = forceType;
 
-  Property.string({required String name, required String value})
+  Property.string(
+      {required String name, required String value, PropertyType? forceType})
       : _name = name,
-        _value = value;
+        _value = value,
+        _forceType = forceType;
 
-  Property.date({required String name, required DateTime value})
+  Property.date(
+      {required String name, required DateTime value, PropertyType? forceType})
       : _name = name,
-        _value = value;
+        _value = value,
+        _forceType = forceType;
 
-  Property.intArray({required String name, required List<int> value})
+  Property.intArray(
+      {required String name, required List<int> value, PropertyType? forceType})
       : _name = name,
-        _value = value;
+        _value = value,
+        _forceType = forceType;
 
-  Property.doubleArray({required String name, required List<double> value})
+  Property.doubleArray(
+      {required String name,
+      required List<double> value,
+      PropertyType? forceType})
       : _name = name,
-        _value = value;
+        _value = value,
+        _forceType = forceType;
 
-  Property.stringArray({required String name, required List<String> value})
+  Property.stringArray(
+      {required String name,
+      required List<String> value,
+      PropertyType? forceType})
       : _name = name,
-        _value = value;
+        _value = value,
+        _forceType = forceType;
 }
 
 class Event {
@@ -59,7 +71,13 @@ class Event {
     return {
       "name": _name,
       "data": _properties != null
-          ? {for (var property in _properties) property._name: property._value}
+          ? {
+              for (var property in _properties)
+                property._name: {
+                  "value": property._value,
+                  "forceType": property._forceType
+                }
+            }
           : null
     };
   }
@@ -77,9 +95,7 @@ class PianoAnalytics {
   PianoAnalytics(
       {required int site,
       required String collectDomain,
-      PianoAnalyticsVisitorIDType visitorIDType =
-          PianoAnalyticsVisitorIDType.uuid,
-      PianoConsents? consents,
+      VisitorIDType visitorIDType = VisitorIDType.uuid,
       MethodChannel? channel})
       : _parameters = {
           "site": site,
@@ -97,6 +113,26 @@ class PianoAnalytics {
     _checkInit();
     await _channel.invokeMethod(
         "send", {"events": events.map((event) => event.toMap()).toList()});
+  }
+
+  Future<void> privacyIncludeStorageFeatures(
+      {required List<PrivacyStorageFeature> features,
+      required List<PrivacyMode> modes}) async {
+    _checkInit();
+    await _channel.invokeMethod("privacyIncludeStorageFeatures", {
+      "features": features.map((feature) => feature.value).toList(),
+      "modes": modes.map((mode) => mode.value).toList()
+    });
+  }
+
+  Future<void> privacyExcludeStorageFeatures(
+      {required List<PrivacyStorageFeature> features,
+      required List<PrivacyMode> modes}) async {
+    _checkInit();
+    await _channel.invokeMethod("privacyExcludeStorageFeatures", {
+      "features": features.map((feature) => feature.value).toList(),
+      "modes": modes.map((mode) => mode.value).toList()
+    });
   }
 
   void _checkInit() {