+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.1
+* PianoConsents has been moved to the piano_consents package
+* Added feature to specify property type via forceType argument
+
 ## 1.0.0
 
 * Added plugins for PianoAnalytics and PianoConsents